Use the existing component
Also move over the duplicate code to DRY things out a bit.
