+++ src/templates/common/phoneNumber/index.tsx
@@ -1,22 +1,39 @@
 import { PhoneNumber as FormattedPhoneNumber } from '@/types/formatted/phoneNumber'
 
-const separateExtension = (number: string) => {
+export interface SeparatedPhoneNumber {
+  phoneNumber: string
+  extension: string
+  processed: boolean
+}
+
+export const separateExtension = (
+  phoneNumber: string
+): SeparatedPhoneNumber | null => {
+  if (!phoneNumber) {
+    return null
+  }
+
   const phoneRegex =
     /\(?(\d{3})\)?[- ]*(\d{3})[- ]*(\d{4}),?(?: ?x\.? ?(\d*)| ?ext\.? ?(\d*))?(?!([^<]*>)|(((?!<v?a).)*<\/v?a.*>))/i
-  const match = phoneRegex.exec(number)
+
+  const match = phoneRegex.exec(phoneNumber)
 
   if (!match || !match[1] || !match[2] || !match[3]) {
-    // Short number or not a normal format
-    return { number, extension: '' }
+    // Short number, invalid format, or not a full US number
+    return {
+      phoneNumber,
+      extension: '',
+      processed: false,
+    }
   }
 
   const phone = `${match[1]}-${match[2]}-${match[3]}`
-  // optional extension matching x1234 (match 4) or ext1234 (match 5)
   const extension = match[4] || match[5] || ''
 
   return {
     phoneNumber: phone,
     extension,
+    processed: true,
   }
 }
 
@@ -55,7 +72,7 @@ export const PhoneNumber = (
   }
 
   return (
-    <div className={className || undefined}>
+    <p className={className || undefined}>
       <strong>{labelToDisplay}: </strong>
       <va-telephone
         contact={numberToDisplay.replace?.(/-/g, '')}
@@ -65,6 +82,6 @@ export const PhoneNumber = (
         sms={sms ? true : undefined}
         tty={tty ? true : undefined}
       />
-    </div>
+    </p>
   )
 }
